Update soap dependency
No breaking changes are documented from 0.43 to 0.45, but this is a breaking
change for consumers of node-soap-graphql because they need to update the
soap dependency. We could probably allow the full range, but as we are
about to release a new major version anyway, this is a good time to bump
the peer dependency.

A change in 0.44 could be considered breaking: axios is now a peer
dependency. To make this more apparent, we declare it as a peer dependency
ourselves.